Move TimeProvider into dd-sdk-android-internal
