+++ dd-sdk-android-core/src/test/kotlin/com/datadog/android/core/internal/CoreFeatureTest.kt
@@ -32,11 +32,11 @@ import com.datadog.android.core.internal.system.NoOpSystemInfoProvider
 import com.datadog.android.core.internal.thread.BackPressuredBlockingQueue
 import com.datadog.android.core.internal.time.AppStartTimeProvider
 import com.datadog.android.core.internal.time.KronosTimeProvider
-import com.datadog.android.core.internal.time.NoOpTimeProvider
 import com.datadog.android.core.internal.user.DatadogUserInfoProvider
 import com.datadog.android.core.internal.user.NoOpMutableUserInfoProvider
 import com.datadog.android.core.persistence.PersistenceStrategy
 import com.datadog.android.core.thread.FlushableExecutorService
+import com.datadog.android.internal.time.DefaultTimeProvider
 import com.datadog.android.ndk.internal.DatadogNdkCrashHandler
 import com.datadog.android.ndk.internal.NoOpNdkCrashHandler
 import com.datadog.android.privacy.TrackingConsent
@@ -1309,7 +1309,7 @@ internal class CoreFeatureTest {
         assertThat(testedFeature.systemInfoProvider)
             .isInstanceOf(NoOpSystemInfoProvider::class.java)
         assertThat(testedFeature.timeProvider)
-            .isInstanceOf(NoOpTimeProvider::class.java)
+            .isInstanceOf(DefaultTimeProvider::class.java)
         assertThat(testedFeature.trackingConsentProvider)
             .isInstanceOf(NoOpConsentProvider::class.java)
         assertThat(testedFeature.userInfoProvider)
