+++ features/dd-sdk-android-session-replay/src/main/kotlin/com/datadog/android/sessionreplay/internal/DefaultRecorderProvider.kt
@@ -22,6 +22,7 @@ import android.widget.TextView
 import androidx.appcompat.widget.ActionBarContainer
 import androidx.appcompat.widget.SwitchCompat
 import com.datadog.android.api.feature.FeatureSdkCore
+import com.datadog.android.internal.time.DefaultTimeProvider
 import com.datadog.android.internal.utils.ImageViewUtils
 import com.datadog.android.sessionreplay.ImagePrivacy
 import com.datadog.android.sessionreplay.MapperTypeWrapper
@@ -85,7 +86,7 @@ internal class DefaultRecorderProvider(
             touchPrivacyManager = touchPrivacyManager,
             textAndInputPrivacy = textAndInputPrivacy,
             recordWriter = recordWriter,
-            timeProvider = { System.currentTimeMillis() },
+            timeProvider = DefaultTimeProvider(),
             mappers = customMappers + builtInMappers(),
             customOptionSelectorDetectors = customOptionSelectorDetectors,
             customDrawableMappers = customDrawableMappers,
