--- features/dd-sdk-android-session-replay/src/main/kotlin/com/datadog/android/sessionreplay/internal/utils/TimeProvider.kt
@@ -1,16 +1,0 @@
-/*
- * Unless explicitly stated otherwise all files in this repository are licensed under the Apache License Version 2.0.
- * This product includes software developed at Datadog (https://www.datadoghq.com/).
- * Copyright 2016-Present Datadog, Inc.
- */
-
-package com.datadog.android.sessionreplay.internal.utils
-
-/**
- * Provides the current device timestamp.
- * Meant for internal usage.
- */
-internal fun interface TimeProvider {
-
-    fun getDeviceTimestamp(): Long
-}
